requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
drf-spectacular>=0.27,<0.29
python-dotenv>=1.0,<2
PyMySQL>=1.1,<2
